React three fiber rendering experiments (#1)
* Adding react-three-fiber

* Fix eslint warning

* Stuff kinda workin in 3d!

* Fix duplicate keying bug on restart

* threejs types

* eslint tweak

* Trying out a much bigger map texture asset

* Some hacky stuff to make the camera manageble for now...

Far from complete, I will need to do a large deep dive into this system
and really understand how to build a better version, but all of this is
pretty new to me

* Adding zustand to the project

* Adding foundational code for zustand store transition

* Improve bun support for server

* Update to new gzipped format

* Port useWebSocket over to zustand store

Mostly prep work for thinking about data muxing

* Needless whitespace change

* Phase 1 of data processing

Alright, I actually think this might be a good starting place to mimic
some of the squadron/path handling.  This should potentially be enough
to start playing with a renderer...

* Tracking the big boi in lfs

* Updating to some new log data

* Misc server fixes

* Fix unitId parsing

* Add react spring dep

* Starting to get a functional map interface working

* Refactored arrows with an actual model

Holy jeez tonight was a lot of head banging against the wall.  It was
incredibly difficult to get this all working with all the issues I ran
into, but we have something, and it kinda works, which is nice...

* Server tweak

Reduce the tick rate to 3 updates (emulating once per 3 minutes).  Also
add more helpful json parsing

* Improve error handling of websocket stuff

Don't make it restart, and instead just skip the update, because...

* Fix major bug on point cleanup

loooll

* Add a debug mode toggle and debug visualization to markers/paths

* Experimenting with new translation code...

Probably still not correct, but a good stopping place for tonight...

* Host the server

* Blender 3d model stuffs

* Adding some simple models
* Also added a script command to generate the components

* First pass at decently rendering block marker info

Also took the time to refactor and organize things a bit more

* First pass at a clock and colors implementation

* Quick pass at a new custom arrow model

* Add types to geometry loaded stuff

* Tweak Arrow props

* Blender stuff... I guess?

* Using a more compressed asset for the map... for now